KAFKA-3859: Fix describe group command to report valid status when group is empty
With the new consumer, when all consumers of a consumer group are stopped (i.e. the group is empty), the describe consumer group command returns `Consumer group ... is rebalancing.`
This PR fixes this issue by distinguishing between when the consumer group is actually rebalancing and when it has no active members.