MINOR: Remove slf4j-log4j from kafka-streams compile dependencies
As kafka-streams is intended to be used by applications that may or may not wish to use log4j, kafka-streams itself should not have a dependency on a concrete log framework.  This change adapts the dependencies to be API-only for compile, and framework-specific for the test runtime only.

I read through the [Contributing Code Guidelines](https://cwiki.apache.org/confluence/display/KAFKA/Contributing+Code+Changes) and interpreted this as a trivial change that doesn't require a Jira ticket.  Please let me know if I've interpreted that wrongly.

This contribution is my original work and I license the work to the project under the project's open source license.

Author: Mathieu Fenniak <mathieu@encouragemarketing.com>

Reviewers: Guozhang Wang <wangguoz@gmail.com>

Closes #1639 from mfenniak/fix-slf4j-dependency-for-streams

(cherry picked from commit cfebfdfa5e537946e4c6e1cc986f20131899be31)
Signed-off-by: Guozhang Wang <wangguoz@gmail.com>