MINOR: Small enhancement to Deserializer Javadoc
I’ve implemented my own custom Deserializer and been using it with `KStream.reduceByKey`; I observed that `reduceByKey` was passing null to my implementation, but it wasn’t clear to me what my implementation was expected to do in this case. So this attempts to clarify it.

This is my original work and I license this work to the Kafka project under Kafka’s open source license (the Apache License 2.0).

Author: Avi Flax <avi@aviflax.com>

Reviewers: Eno Thereska <eno.thereska@gmail.com>, Ismael Juma <ismael@juma.me.uk>, Guozhang Wang <wangguoz@gmail.com>, Ewen Cheslack-Postava <ewen@confluent.io>

Closes #1503 from aviflax/patch-1