KAFKA-4018; Streams causing older slf4j-log4j library to be packaged along with newer version
This is a regression caused by 0bb1d3ae.

After that commit, Streams no longer has a direct dependency on slf4j-log4j12, but zkclient
has a dependency on an older version of slf4j-log4j12, so we get a transitive dependency on
the older version.

The fix is to simply exclude the undesired dependencies from the zkclient dependency.

Author: Ismael Juma <ismael@juma.me.uk>

Reviewers: Guozhang Wang <wangguoz@gmail.com>

Closes #1704 from ijuma/kafka-4018-streams-duplicate-slf4j-log4j