KAFKA-3817: KTableRepartitionMap publish old Change first, for non-count aggregates
I affirm that the contribution is my original work and that I license the work to the project under the project's open source license.

This cleans up misbehaviour that was introduce while fixing KAFKA-3817. It is impossible for a non-count aggregate to be build, when the addition happens before the removal. IMHO making sure that these details are correct is very important.

This PR has local test errors. It somehow fails the ResetIntegrationTest. It doesn't quite appear to me why but it looks like this PR breaks it, especially because the error appears with the ordering of the events. Still I am unable to find where I could have broken it. Maybe not seems to fail on trunk aswell.

Author: jfilipiak <Jan.Filipiak@trivago.com>

Reviewers: Guozhang Wang <wangguoz@gmail.com>

Closes #1705 from Kaiserchen/KAFKA-3817-preserve-order-for-aggreagators

(cherry picked from 3dafb81da788294d4c2e9811f49437608e5b9ce8)