KAFKA-2932: Adjust importance level of Kafka Connect configs
ewencp I went down the list of connect configs and it looks like only the internal converter configs are mismarked.  It looks like the `cluster` config that is present in the current docs is already gone.  The only other values I can see arguing to change importance on are the ssl configs (marked high) but they are consistent with the producer/consumer config docs so that's at least consistent.  Everything else marked high looks either mandatory or requires consideration in a production deployment to me.

Author: Dustin Cote <dustin@confluent.io>

Reviewers: Ewen Cheslack-Postava <ewen@confluent.io>

Closes #1653 from cotedm/KAFKA-2932