KAFKA-3894: allow partial log cleans
As discussed in https://issues.apache.org/jira/browse/KAFKA-3894, this
PR makes the log cleaner do a "partial" clean on a segment, whereby it
builds a partial offset map up to a particular offset in a segment. Once
cleaning resumes again, we will continue from the next dirty offset,
which can now be located in the middle of a segment.

Prior to this PR, segments with overly numerous keys could crash the log
cleaner thread, as it was required that the log cleaner had to fit at
least a single segment in the offset map.