Added the first set of system tests for replication quotas. These tests validate throttling behavior during partition reassigment.
Along with this patch are fixes to the test framework which include:

  1. KakfaService.verify_replica_reassignment: this method was a no-op
and would always return success, as explained in KAFKA-4204. This patch
adds a workaround to the problems mentioned there, by grepping correctly
for success, failure, and 'in progress' states of partition
reassignment.

  2. ProduceConsumeValidateTest.annotate_missing_messages would call
missing.pop() to enumerate the first 20 missing messages. This meant
that all future counts of what is actually missing would be off by 20,
leading to the impression of data loss.