KAFKA-4214; kafka-reassign-partitions fails all the time when brokers are bounced during reassignment
There is a corner case bug, where during partition reassignment, if the
controller and a broker receiving a new replica are bounced at the same
time, the partition reassignment is failed.

The cause of this bug is a block of code in the KafkaController which
fails the reassignment if the aliveNewReplicas != newReplicas, ie. if
some of the new replicas are offline at the time a controller fails
over.

The fix is to have the controller listen for ISR change events even for
new replicas which are not alive when the controller boots up. Once the
said replicas come online, they will be in the ISR set, and the new
controller will detect this, and then mark the reassignment as
successful.

Interestingly, the block of code in question was introduced in
KAFKA-990, where a concern about this exact scenario was raised :)

This bug was revealed in the system tests in https://github.com/apache/kafka/pull/1904.
The relevant tests will be enabled in either this or a followup PR when PR-1904 is merged.

Thanks to junrao identifying the issue and providing the patch.

Author: Apurva Mehta <apurva.1618@gmail.com>

Reviewers: Jun Rao <junrao@gmail.com>

Closes #1910 from apurvam/KAFKA-4214

(cherry picked from commit be6056abc9970600347c95c4c8658799b76dbe6b)
Signed-off-by: Jun Rao <junrao@gmail.com>