KAFKA-4213; System tests for replication throttling (KIP-73)
In this patch, we test `kafka-reassign-partitions` when throttling is active.

This patch also fixes the following:

  1. KafkaService.verify_reassign_partitions did not check whether
partition reassignment actually completed successfully (KAFKA-4204).
This patch works around those shortcomings so that we get the right
signal from this method.

  2. ProduceConsumeValidateTest.annotate_missing_messages would call
`pop' on the list of missing messages, causing downstream methods to get
incomplete data. We fix that in this patch as well.

Author: Apurva Mehta <apurva.1618@gmail.com>

Reviewers: Geoff Anderson <geoff@confluent.io>, Ben Stopford <benstopford@gmail.com>, Ismael Juma <ismael@juma.me.uk>

Closes #1904 from apurvam/throttling-tests

(cherry picked from commit 5a989d1748426004212d44abad4567f9034a04ef)
Signed-off-by: Ismael Juma <ismael@juma.me.uk>