MINOR: Clarify doc on consumption of topics
In doc it stays:

_"Our topic is divided into a set of totally ordered partitions, each of which is consumed by one consumer at any given time."_

And consumer is described as:

_"We'll call **processes** that subscribe to topics and process the feed of published messages **consumers**."_

Which might lead to a wrong conclusion - that each partition can be read by one process at any given time.

I think this statements misses information about **consumer groups**, so i propose:

_"Our topic is divided into a set of totally ordered partitions, each of which is consumed by exactly one consumer (from each subscribed consumer groups) at any given time"_

This contribution is my original work and I license the work to the project under the project's open source license.

Author: pilo <jakub.pilimon@4finance.com>

Reviewers: Jiangjie Qin <becket.qin@gmail.com>, Jason Gustafson <jason@confluent.io>

Closes #1900 from pilloPl/minor/doc-fix