KAFKA-4284: Make Partitioner a Closeable and close it when closing the producer
[KAFKA-4284](https://issues.apache.org/jira/browse/KAFKA-4284)

Even though Partitioner has a close method it is not closed when the producer is closed. Serializers, interceptors and metrics are all closed, so partitioners should be closed to.

To be able to use the same mechanism to close the partitioner as the serializers, etc. I had to make the `Partitioner` interface extend `Closeable`. Since this doesn't change the interface that feels ok and should be backwards compatible.

Looking at [KAFKA-2091](https://issues.apache.org/jira/browse/KAFKA-2091) (d6c45c70fb9773043766446e88370db9709e7995) that introduced the `Partitioner` interface it looks like the intention was that the producer should close the partitioner.

This contribution is my original work and I license the work to the project under the project's open source license.

Author: Theo <theo@iconara.net>

Reviewers: Ewen Cheslack-Postava <ewen@confluent.io>

Closes #2000 from iconara/kafka-4284