MINOR: Clarify log deletion configuration options in server.properties
I spent a bit of time tracking down why files were being deleted before they reached log.retention.hours of age. It turns out that the time and size log retention schemes function independently, and not as the original comment "The minimum age of a log file to be eligible for deletion" might indicate to a new user.

Author: Mark Rose <markrose@markrose.ca>

Reviewers: Guozhang Wang <wangguoz@gmail.com>

Closes #28 from MarkRose/fix_misleading_configuration_file_for_trunk