Some fixes to the system which handles mismatches between client versions and broker versions
If a message cannot be sent to a broker, because it is too old, an
exception will be thrown from AbstractRequest#Builder#build.  This
exception will be picked up in NetworkClient and used to create a
ClientResponse.  We now include the whole exception in the
ClientResponse, for extra detail and clarity in case it is needed.

In the Fetcher, we don't need to check for negative timestamps when
requireTimestamps is set, since this will be caught by the builder's
build method.  This check also was not reliable since sometimes no
offsets at all would be returned, and it would appear as if the broker
supported timestamps, when it did not.