KAFKA-4668: MirrorMaker replicate topics from `earliest`
MirrorMaker currently inherits the default value for
`auto.offset.reset`, which is `latest`.

While for most consumers this is a sensible default, MirrorMakers are
specifically designed for replication, so they should default to
replicating topics from the beginning.

A specific scenario where this really matters is when a MirrorMaker is
subscribed to a regex pattern. If auto-topic creation is enabled on the
cluster, and you start producing to a non-existent topic that matches
the regex, then there will be a period of time where the producer is
producing before the new topic's partitions have been picked up by the
MirrorMaker. Those messages will never be consumed by the MirrorMaker
because it will start from latest, ignoring those just-produced messages.